requirements/production.txt
django==2.2.2
djangorestframework==3.9.4
django-filter==2.1.0
django-environ==0.4.5
psycopg2==2.8.2
pendulum==2.0.4

# Deploy requirement
sentry-sdk==0.9.0
gunicorn==19.9.0
